'pip show {NAME}' also shows the list of recursive dependencies of package '{NAME}'.
