+++ pip/commands/show.py
@@ -2,6 +2,7 @@ import os
 
 from pip.basecommand import Command
 from pip.log import logger
+from pip.util import recursive_dependencies
 from pip._vendor import pkg_resources
 
 
@@ -78,6 +79,8 @@ def print_results(distributions, list_all_files):
         logger.notify("Version: %s" % dist['version'])
         logger.notify("Location: %s" % dist['location'])
         logger.notify("Requires: %s" % ', '.join(dist['requires']))
+        logger.notify("Requires recursive: %s" % ', '.join(
+            recursive_dependencies([dist['name']])))
         if list_all_files:
             logger.notify("Files:")
             if dist['files'] is not None:
